Allowing enable for stack backtracing printouts with compilers other than GCC
Here, I have now allowed the setting of
Teuchos_ENABLE_STACKTRACE:BOOL=ON when GCC is not the compiler but the
default is set based on if the compiler is GCC.  The cmake varible
still will not be defined if BinUtils is not enabled.

This should not break any backward compatibility at all.