Adding XHOST_TEST and XHOSTTYPE_TEST arguments to PACKAGE_ADD_EXECUTABLE_AND_TEST(...)
With this commit, you can now pass in extra arguments to exclude running a
test on XHOST and XHOSTTYPE platforms but not exclude building the test code.
This will allow me to disable some tests for AIX from running but still build
them (which is better than nothing).

This is fully unit tested.