MueLu UnitTest: can now be run with Epetra or Tpetra (using --linAlgebra=...)
Cthulhu parameter '--linAlgebra=' is added to the Teuchos::CommandLineProcessor used by Teuchos::UnitTestRepository::runUnitTestsFromMain.
I use a static Cthulhu::Parameters object stored in the static class MueLu::UnitTest::Parameters. So it is now possible to access the cthulhuParameter::getLib() methods anywhere on the unit test.
I updated the CMakeList.txt of UnitTest and 'make test' runs tests for both Epetra and Tpetra.

In addition, I renamed the namespace MueLu::UnitTest->MueLu::TestHelpers to be consistent with the name of the source file. Functions for creating Map and Matrix are now parts of the static class TestHelpers::Factory.
To resume :
- MueLu::UnitTest::create_1d_poisson_matrix<SC,LO,GO,NO,LMO> is now MueLu::TestHelpers::Factory<SC, LO, GO, NO, LMO>::Build1DPoisson()
- MueLu::UnitTest::getDefaultComm() is now MueLu::TestHelpers::Parameters::getDefaultComm().

Other minor changes:
- renamed main.cpp -> UnitTests.cpp
- renamed test_helpers.hpp -> MueLu_TestHelpers.hpp and MueLu::UnitTest -> MueLu::TestHelpers

- Cthulhu::Parameters and Gallery::Parameters have now an additional setCLP() method.
- removed Epetra dependency in Hierarchy unit test
- renamed cmdp->clp everywhere to be consistent with Teuchos naming