Enable backwards compatibility testing
Enabling some preliminary backwards compatibility testing that is working off of
a dummy "10.8" branch. This is really just to make sure that the backwards
compatibility testing is working and the dummy branch will not be used for
anything other than this testing. The dummy branch only exists on the nightly
repository so it should not affect anything but the test machines.

Once 10.8 is created I will need to replace the dummy branch with the correct
branch for 10.8, but that is a trivial change