Major refactoring of setting compiler flags with good unit tests.
Here a major cleanup of the way that the Trilinos CMake build system sets
compiler flags has been performed and good unit tests have been written.  For
the most part, the behavior of the system has not radically changed but one
major change is that now setting -DTrilinos_ENABLE_CHECKED_STL=ON will result
in having -D_GLIBCXX_DEBUG for all build types for GNU.  Before, it was only
set for CMAKE_BUILD_TYPE=DEBUG.

The motivation for this refactoring and change and the creation of this
significant set of unit tests was because it was discovered that that
MPI_DEBUG build on the checkin-test.py script was not setting up to use the
checked STL.  This allowed some people to push code with checkin-test.py that
was invalid STL code.

There may be some SS code (e.g. STK) that will fail now when when run through
the checkin-test.py script that will need to be fixed.  However, I am pushing
this change because it is important to get this fixed.