Removing LIME from the main Trilinos CI and Nighlty builds.
The current CDash server on trilinos.sandia.gov does not handle extra
labels with add-on packages correctly.  This results in inconsistent
error notification emails and problems with orginzation in the CDash
queries for subprojects.  Hopefully this will be fixed with the
updated version of CDash on testing.sandia.gov but no idea when that
will go online.

Also the current LIME devleopers appear to be ignoring failure emails
so there is little point in keeping the CI and Nightly builds of LIME
on trilinos-test2 going.  We will have a CASL-specific LIME build here
shortly which will be rigorously maintained.