Extended NOX::Epetra::LinearSystemStratimikos for transpose solves.
In applying LOCA to the CSTR equations in Albany it became apparent that
NOX::Epetra::LinearSystemStratimikos doesn't work with the minimally
augmented bifurcation methods.  The issue was setting the preconditioner
operator, which LOCA does to set preconditioners for transpose linear
systems.  This commit adds support for setting the preconditioner
operator that seems to work and adds a test that uses the minimally
augmented turning point and stratimikos.