(From Sierra) Optimizations for FEI linear-system assembly.
For Aztec solver blocks, there was an un-necessary communication happening in
the case where the no-column-mod option is on. (If no column-mod is done for
Dirichlet BCs, then BC data doesn't need to be gathered from neighboring processors.)

For Trilinos solver blocks, which assemble into Epetra matrices, an optimization
was added which makes better use of local-equation-ids where possible (rather
than always passing global-ids to Epetra). This allows Epetra to make far fewer
calls to the Epetra_Map::LID methods.