Adding new generic-looping-demon.py driver for CI processes (bug 4681)
The new python script generic-looping-demon.py will run an arbitrary command
in a loop, sleeping a given amount of time between iterations, and stop by a
set absolute time.

I wrote this script to drive a CASL process to move commits from SSG to an
internal ORNL computer but this same script will also be used to set up a CI
server for Trilinos and CASL add-on packages on an ORNL machine.  This same
script should be used to refactor the existing Trilinos CI process and will
fix a bunch of issues with the current driver (I will comment in bug 4681).

This script is not currently unit tested.  However, it is quite small, it is
easy to manually, and I don't see anyone needing to change much of anything in
the future (except for setting a relative time limit or give an option to just
set the time in the current date).

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=86,notpassed=0 (1.05 min)
1) SERIAL_RELEASE => passed: passed=86,notpassed=0 (0.86 min)