Adding in cheaper diagonal scaling.
Commiting two new utilities:  GetMatrixDiagonal and ScaleMatrix.  GetMatrixDiagonal returns the diagonal in an ArrayRCP.
ScaleMatrix left scales a matrix by (the inverse of) a diagonal matrix in place.  Using the inverse is the diagonal.
Haven't enabled an Epetra version yet, so this will throw right now.

This lets us get rid of both the D and DAP operators in the SaPFactory.