Fixing behavior of script for determining modified files after last refactor
Turns out I broke the CI integration script for CASL when Dakota was added.
Now it should work but it currently ignores changes to Dakota that would
trigger a rebuild of TriKota.  This can be fixed but it will take some work
and TriKota is not being CI tested yet so it is not needed yet.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=86,notpassed=0 (1.15 min)
1) SERIAL_RELEASE => passed: passed=86,notpassed=0 (1.04 min)
Other local commits for this build/test group: d679d88