(From Sierra) Add ability to swap memory allocators
Add new bjam feature, allocator=native|ptmalloc3, that can be set
to ptmalloc3 to use our modified version of the ptmalloc3 package.
This can be extended to include other allocators like tcmalloc in
the future.

--
The ptmalloc3 allocator in TPLs_src/ptmalloc3 has been modified to
include a new

    size_t malloc_used()

call which returns the amount of used memory, including overhead and
alignment.  This call walks the arena memory (not every allocation,
just the arenas), so its cheap but not trivial.

--
Added a simple unit test which does what it can to check
functionality, but since the deallocation does not necessary return
all the memory the before allocation deallocation does not always
equal the value prior to allocation.

--
Fixed "hidden" variable issue in Fmwk_MeshObjConnect.C