Basic subpackage support (bug 4644)
This commit represents basic subpackage support.  I am not totally happy with
the increased complexity that this change has had on the PackageArch system
but there is minimal extra complexity and overhead needed to define
subpackages with this system.

There are decent unit tests for the addition of subpackages but more should be
added perhaps.

The example of how to set up subpackages will be the Thyra package (to be
committed next).