Fix more dependencies on Thyra adapters.
I believe that I have fixed the majority of the unstated/unprotected
dependencies on the Thyra adapters with this commit. Basically I turned on Thyra
and ThyraCore and disabled the Thyra adapters explicitly and turned on forward
dependencies and fixed what issues came up. There were a number of issues that
came up that weren't necessarily from a dependence on Thyra adapters, for
instance a number of stratimikos dependencies failed, but only because
stratimikos was disabled because of a required dependency on the Thyra adapters.
It is possible that there is still some code that has a dependecy that is an
option or requires a tpl I don't have etc. but I believe this is the majority of
the issues we've been bumping into.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, FEApp, Anasazi, Piro, Rythmos, Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=414,notpassed=0 (162.04 min)
1) SERIAL_RELEASE => passed: passed=417,notpassed=0 (78.14 min)