Add implementations of tgamma functions
Intrepid recently started using the tgamma functions which are part of the posix
and c99 standard for cmath that are apparently not available on windows. I have
put in a lanczos approximation of the gamma function for windows only. This
approximation is not 100% equivalent to the tgamma output, but it is generally
pretty close with a relative diff of less than 2.0e-13 for the double version of
tgamma.