Partial refactor of TentativePFactory.
Reduces memory usage in QR:
  Rows are now inserted upon completion of each aggregate-wise QR.
  This eliminates the temporary CSR arrays, which effectively doubled the memory usage.

Note: this also changes the row map that Ptent uses back to one compatible with the aggregate ownership,
instead of using the row map of the fine level A matrix.  This is not compatible with Epetra, as you
can't insert rows off-processor.

Next steps: Store off-processor rows of Q separately,
            Ship them off to the owning processors, and let those processors insert them.
            This is basically doing explicitly what Tpetra must be doing implicitly.