Fix Moertel configuration error
Disabling the tpl exodus yesterday removed some env variables that were allowing
netcdf to be found in cases where the path to it wasn't explicitly enabled. I am
adding the path to the gcc built netcdf to the pgi build since it is needed for
the seacas exodus and had been working previously.