Adding configure option Epetra_ENABLE_Fortran that defaults to ${PROJECT_NAME}_ENABLE_FORTRAN
This change was made to allow selectively disabling Fortran support in
Epetra when it is on in the rest of Trilinos.  This was to get around
an aggressive Intel Fortran compilier optimization that was causing
the test EpetraExt_inout_test to fail.

The default behavior is exactly the same, when
${PROJECT_NAME}_ENABLE_FORTRAN=ON, Epetra builds Fortran files for key
kernals.  However, if you set Epetra_ENABLE_Fortran=OFF you get the
new behavior.