MueLu: adapt unit tests to changes from last commit.
Note: If we have some data without a generating factory we need a
default Factory handler. The default factory handler is usually generated
in the constructor of the Hierarchy class. So, when using
Level.Set("A",A) with the default factory, the level class first has to be
added to the Hierarchy (by calling Hierarchy.SetLevel(...))!