MueLu: added NoFactory
NoFactory is a singleton class, that provides a dummy factory pointer for
user provided (user defined) variables and data to be stored in the level class.
It's intended to be used only in the user driver routines for setting the variables
of the finest level!
The Level Get/Set/Request/Release routines have been adapted accordingly, such that
the variants without a generating factory pointer automatically use the
MueLu::NoFactory::get() pointer.

Note: it is
lvl.Set('A') != lvl.Set('A',NULL)
but
lvl.Set('A') == lvl.Set('A', MueLu::NoFactory::get()).