Enabling outer package if any subpackages are enabled; cleaning up enbale printouts
With this commit, when any subpackage is enabled, the outer package
enable will also be set.  This makes the output look more logical
(e.g. if the subpackage ThyraCore is enabled then so is outer package
Thyra) but also helps clean up some logic, such as in the
TrilinosCTestDriverCore.cmake script.

In the case of the CTest driver, when only a subpackage gets enabled,
now the outer package will also get enabled and this will cause the
package to get processed in the CTest/CDash testing.  Before, it was
not and this was causing one of the CASL-related packages (CASLRAVE)
to not get processed correctly.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=89,notpassed=0 (1.95 min)
1) SERIAL_RELEASE => passed: passed=89,notpassed=0 (1.14 min)
2) MPI_DEBUG_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
3) SERIAL_RELEASE_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
Other local commits for this build/test group: 481854f