Ensure that CMAKE_CURRENT_LIST_DIR is usable from the install tree.
CMake versions < 2.8.3 do not have this variable. This commit
adds a code snippet to the beginning of all the *Config.cmake
files that ensures this variable is defined within the install
tree. This is to make sure the Trilinos install tree is valid
even when it is re-rooted at install time by packaging systems
that support this option.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => passed: passed=89,notpassed=0 (3.76 min)