Adding CMAKE_${LANG}_FLAGS_${BUILDTYPE}_OVERRIDE
With the new cmake variable CMAKE_${LANG}_FLAGS_${BUILDTYPE}_OVERRIDE, you can
now override more selectively the debug or release optimizaition levels for a
given compiler.  This is needed to allow, for example, to do a
CMAKE_BUILD_TYPE=DEBUG but use -O3 for Fortran (which is the case for CASL).