Fix bug in CrsMatrix
This addresses bug 5370. A memory deallocation error that Martin Kronbichler
reported. A previous fix was applied for this, however that fix caused issues
with Teko. This change fixes the issue that Martin was seeing in a better way
and does not cause Teko to fail.

Build/Test Cases Summary
Enabled Packages: Epetra
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=779,notpassed=0 (92.20 min)
1) SERIAL_RELEASE => passed: passed=753,notpassed=0 (34.82 min)
Other local commits for this build/test group: 2300873