Fixing crash in TentativePFacory for index != 1.
For a global ID not on the current processor, Tpetra returns ScalarType::invalid, but Epetra returns index-1.
This is consistent for index=0, as OrdinalTraits<>::invalid casts to -1, which explains why we often don't see this.
For index not equal to 1, however, the check that a global row is locally owned was passing (incorrectly) in
the QR loop.  As a result, Epetra was trying to insert global rows that were non-local.  (Tpetra can do this just
fine.) The fix is simply to return -1 whenever the GID is not found.

Summary of changes (lots)
-------------------------
Xpetra_EpetraMap.* specialization for getGlobalElement
  main point of this commit
MueLuTentativePFactory.hpp
  Changing index to GO.
unit_tests/TentativePFactory.cpp
  Enabling test nonstandard maps test.

Xpetra_EpetraMap.*    specialization for getGlobalElement
  this was previously hand-coded
Xpetra_EpetraMultiVector.* specialization for setSeed and randomize
  this was previously hand-coded
Xpetra_EpetraVector.* specialization for randomize
  this was previously hand-coded
Xpetra_TpetraCrsMatrix.*
  specialization getLocalDiagCopy
  add typedef for TpetraVectorClass
  these were previously hand-coded
Xpetra_TpetraMultiVector.* specialization for randomize and elementWiseMultiply
  this was previously hand-coded
Xpetra_EpetraCrsGraph.cpp
  Implementing describe method.
Xpetra_TpetraCrsGraph.hpp
  Removed some unnecessary headers:  ?petra_Utils.hpp, ?petra_TpetraMap.hpp, ?petra_TpetraImport.hpp