Update to exclude system to allow subset distribution tarballs
Changed the behavior of the source distribution exclude system to automatically
exclude any packages that are not enabled. Previously the system would
distribute all packages that weren't explicitly excluded as part of the global
exclude list. This allows much finer control of what will be released in a
tarball at configure time. This also fixes what had been a potential problem
with the package exclude lists. The package exclude lists are only obeyed when
the package is enabled. This is because those exclude lists are listed in the
package's CMakeLists.txt file. If the package wasn't enabled then its exclude
list was never added and therefore wasn't obeyed. Generally for releases this
wasn't an issue since we did enable all released packages for release tarballs.

This required a change to the configure for generating our release tarballs.
Instead of just expecting all the packages to be released we now explicitly
state all packages that are to be released. This will mean that we will have to
update the configure script occasionally, but generally it won't be many
packages that will be added and it should really only happen once per release.
This script also has one hokey thing in it that we fake a Dakota instance to
allow TriKota to configure. This faked Dakota is mostly a noop just to allow a
configure. All of the faked files are removed before a tarball is created so
that they are not mistakenly released.

Also added all the scripts that are used for testing a tarball before releasing
it. These scripts are tied pretty closely to the machine that they are run on,
but they could be modified to work on other machines reasonably easy. This is
mostly to ensure that these scripts aren't just tribal knowledge that one or two
people have.