Fix SEACAS not finding ncdump during configure
SEACAS was looking for ncdump for several of its exodus tests and wasn't finding
it by default  even though netcdf is a required tpl for exodus and therefore in
a known location. I added the path to netcdf to the list of places to look for
ncdump. There is a slight posibility that this won't always work since the
variable I used can store multiple paths and not just one. But since netcdf is
not a very complicated tpl I don't imagine it will ever be the case that it will
hold more than one path. Also since we don't look for the bin directory
explicitly I had to use a relative path from the found library. In the worst
case scenario where one or both of these assumptions are broken though all that
will happen is that ncdump won't be found, which is exactly what happens now.