Allowing update of CMake to fail and not bring down driver
With the shift to not driving CMake development by Trilinos, we are changing
the behavior of this script so that it does not fail hard if CMake can't be
updated.  This issue killed all of the Trilinos testing on a dozen machines
for the last two days (silently).

We also changed from --installer-type=dev to release which never should have
been dev in the first place.

Note that testing of the download-cmake.py script still occurs as a CTest
driven test so if it fails, we will know it.