Changing all paths in PackagesList.cmake and TPLsList.cmake relative to Repository base
With this change, all of the paths listed in the PackagesList.cmake and
TPLsList.cmake are now relative to the Repository base directory for which
these files reside.  This now means that Tribits knows nothing of 'packages'
or 'cmake/TPLs' directories.  The location of packages and TPLs is now
arbitrary and clean.  This helped me to clean up a lot of messed-up logic with
the old approach of blessing 'packages/' and 'cmake/TPLs/'.

As part of this, I changed the convention of TPLsList.cmake file so that you
can just list the path to a find TPL and it will look for the standard name.
There is no great unit testing for specifically naming the file but that can
come later if needed.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Packages
0) MPI_DEBUG => passed: passed=1126,notpassed=0 (13.87 min)
1) SERIAL_RELEASE => passed: passed=1104,notpassed=0 (7.05 min)
2) MPI_DEBUG_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
3) SERIAL_RELEASE_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
Other local commits for this build/test group: 99ceeeb