Fix issue with Teuchos LAPACK on windows shared build
The specializations of the LAPACK class needed to be exported for the windows
shared build.

The default implementations of several of the LAPACK wrappers had void return
values but were trying to return the value from the undefined routine which
caused an error on windows and in general is illegal. The offending routines
have been changed to not return anything, but instead just call the undefined
routine.

Change the type of the undefined class place holder to OrdinalType instead of
ScalarType in ILAENV. The function returns a type of OrdinalType which is not
always compatible with the ScalarType (for instance with complex values). This
will produce a slightly misleading error in the case where the LAPACK class is
instantiated with a type that is not supported, but I think it will still be
clear enough for a developer to fix the issue. ILAENV has to return something
and this is still the cleanest way to deal with that even though it isn't 100%
accurate.

The DLL macro that is shared by all Trilinos packages was expecting the variable
Trilinos_BUILD_SHARED_LIBS to be set if shared libraries are being built. It
looks like the Tribits changes are not setting Trilinos_BUILD_SHARED_LIBS to the
value of BUILD_SHARED_LIBS though. It also looks like the
Trilinos_BUILD_SHARED_LIBS is potentially on the chopping block, which makes
sense as there is no reason to have another variable with the same value but
with the project's name in it when the cmake provided one works perfectly well.
The DLL macro now expects that BUILD_SHARED_LIBS will be set.