Making checkin-test.py enabled all packages when only changing top-level build/test file
I updated the unit tests to show that if you only set --enable-all-pacakges=on
or just change a top-level build/test file that does *not* enable
TrilinosFramework, then the configure would never happen.  Here, I have
updated the unit tests and fixed the code so that now the checkin-test.py
script will propertly enabled all packages and actually run
configure/build/test properly.

This was causing problems for updates of the TriBITS sysytem in
TribitsExampleProject.