Fixed up behavior for support for extra repos
The new behavior eliminates the need for the varaible
${PROJECT_NAME}_SUPPORT_EXTRA_REPOS.  If a project defines the default extra
repos file, it will be used as the default for
${PROJECT_NAME}_EXTRA_REPOSITORIES, otherwise, it must be set by the user to
be used.

This is consistent with the behavior of other optional TriBITS features.  If
the file exists, it will be used.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Packages
0) MPI_DEBUG => passed: passed=1126,notpassed=0 (16.49 min)
1) SERIAL_RELEASE => passed: passed=1104,notpassed=0 (11.88 min)