Setting ${PROJECT_NAME}_ALLOW_NO_PACKAGES=ON for automated package-based CTest testing
By making this one change, now when a package's configure or library build
fails, it will disable that package in all downstream packages but will not
result in errors on those downstream packages (just nothing to build or test).

This is a major change in the system.  With this change, when a funamental
configure failure happens, it will not result in a cascade of 30+ configure
failure messages.  This means, however, that the Trilinos Framework team is
going to need to be more diligent in looking at the dashboard to make sure
that configure failures are taken care of.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Packages
0) MPI_DEBUG => passed: passed=1126,notpassed=0 (5.31 min)
1) SERIAL_RELEASE => passed: passed=1104,notpassed=0 (3.95 min)
Other local commits for this build/test group: 9d8ca87, b614355, 6e33cf5, f204152, 188ffd0