Revering setting Trilinos_ENABLE_ALL_PACKAGES=ON because it is a non-opt to the CTest script
Revert "Adding explicit Trilinos_ENABLE_ALL_PACKAGES=ON for tests that require this"

This reverts commit eea4714829720ef404c14249342cbdd919ad71bf.

That was a silly un-studied change.  The code in TribitsCTestDriverCore.cmake
does not read Trilinos_ENABLE_ALL_PACKAGES=ON from the env and the behavior of
the script is to set this if and only if Trilinos_PACAKGES is emtpy.  The
problem is instead and env cross-talk issue with
CTEST_ENABLE_MODIFIED_PACKAGES_ONLY which I will fix next.