Disable a Teuchos test on windows
The FancyOutput test is not being run on the windows shared build because a
dependent library is not being created as a dll properly. Efforts to try and fix
that library aren't moving forward very fast due to an odd issue with exporting
the vtable for one of the classes. I am disabling this test on windows
temporarily.