Adding unit tests to show inconsistent behavior of zero views of null/empty arrays (bug 5459)
These new unit tests show that ArrayRCP and ArrayView don't allow zero views
of null (empty) arrays.  However, I found a unit test for Array that shows
that it allows zero views of an empty Array object.  That is inconsistent
behavior.

This commit just creates the unit tests that show the inconsistent behavior.
The next commit will add the code to fix the tests.