Migrating changes from modifed version of pamgen to Trilinos version. Changes: Changed bisection decomposition message to: must be divisible by twice the number of trisection_blocks from: must be divisible by the number of trisection blocks.
The number of blocks in each direction must now be explicitly set.
Default was previously 1. it is now 0 and will result in an error.

Fixed parser memory leak.

A fix to the decomposition calculation that changes the modulus calculation
  //The following prevents remainder building up on last cut.
     long long local_split_size = inc_nels[split_direction];

-    if((dels[split_direction]%remaining_cuts[split_direction]) >= 1)local_split_size ++;
+    if((dels[split_direction]%remaining_cuts[split_direction]) > 1)local_split_size ++;
and results in different decompositions.
Several benchmarks in Trilinos needed to be re-baselined because of this new
decomposition. As previously benchmarks were rebaselined in ALEGRA.